Move optimize with rolling horizon method to execute module
Since this function is now not only used in the market_optimization, but also in the temporal disaggregation
