+++ etrago/execute/__init__.py
@@ -387,6 +387,204 @@ def optimize(self):
         print("Method not defined")
 
 
+def optimize_with_rolling_horizon(
+    n,
+    pre_market,
+    snapshots,
+    horizon,
+    overlap,
+    solver_name,
+    extra_functionality,
+    args,
+):
+    """
+    Optimizes the network in a rolling horizon fashion.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    snapshots : list-like
+        Set of snapshots to consider in the optimization. The default is None.
+    horizon : int
+        Number of snapshots to consider in each iteration. Defaults to 100.
+    overlap : int
+        Number of snapshots to overlap between two iterations. Defaults to 0.
+    **kwargs:
+        Keyword argument used by `linopy.Model.solve`, such as `solver_name`,
+
+    Returns
+    -------
+    None
+    """
+    if snapshots is None:
+        snapshots = n.snapshots
+
+    if horizon <= overlap:
+        raise ValueError("overlap must be smaller than horizon")
+
+    # Make sure that quadratic costs as zero and not NaN
+    n.links.marginal_cost_quadratic = 0.0
+
+    starting_points = range(0, len(snapshots), horizon - overlap)
+    for i, start in enumerate(starting_points):
+        end = min(len(snapshots), start + horizon)
+        sns = snapshots[start:end]
+        logger.info(
+            f"""Optimizing network for snapshot horizon
+            [{sns[0]}:{sns[-1]}] ({i+1}/{len(starting_points)})."""
+        )
+
+        if not n.stores.empty:
+            stores_no_dsm = n.stores[
+                ~n.stores.carrier.isin(
+                    [
+                        "PtH2_waste_heat",
+                        "PtH2_O2",
+                        "dsm",
+                        "battery_storage",
+                        "central_heat_store",
+                        "H2_overground",
+                        "CH4",
+                        "H2_underground",
+                    ]
+                )
+            ].index
+            if start != 0:
+                n.stores.loc[stores_no_dsm, "e_initial"] = n.stores_t.e.loc[
+                    snapshots[start - 1], stores_no_dsm
+                ]
+            else:
+                n.stores.loc[stores_no_dsm, "e_initial"] = (
+                    pre_market.stores_t.e.loc[
+                        snapshots[start - 1], stores_no_dsm
+                    ]
+                )
+
+            # Select seasonal stores
+            seasonal_stores = n.stores.index[
+                n.stores.carrier.isin(
+                    [
+                        "central_heat_store",
+                        "H2_overground",
+                        "CH4",
+                        "H2_underground",
+                    ]
+                )
+            ]
+
+            # Set e_initial from pre_market model for seasonal stores
+            n.stores.e_initial[seasonal_stores] = pre_market.stores_t.e.loc[
+                snapshots[start - 1], seasonal_stores
+            ]
+
+            # Set e at the end of the horizon
+            # by setting e_max_pu and e_min_pu
+            n.stores_t.e_max_pu.loc[
+                snapshots[end - 1], seasonal_stores
+            ] = pre_market.stores_t.e.loc[
+                snapshots[end - 1], seasonal_stores
+            ].div(
+                pre_market.stores.e_nom_opt[seasonal_stores]
+            ).clip(
+                lower=0.0
+            ) * (
+                1 + 1e6
+            )
+            n.stores_t.e_min_pu.loc[
+                snapshots[end - 1], seasonal_stores
+            ] = pre_market.stores_t.e.loc[
+                snapshots[end - 1], seasonal_stores
+            ].div(
+                pre_market.stores.e_nom_opt[seasonal_stores]
+            ).clip(
+                lower=0.0
+            ) * (
+                1 - 1e6
+            )
+            n.stores_t.e_min_pu.fillna(0.0, inplace=True)
+            n.stores_t.e_max_pu.fillna(1.0, inplace=True)
+
+        if not n.storage_units.empty:
+            n.storage_units.state_of_charge_initial = (
+                n.storage_units_t.state_of_charge.loc[snapshots[start - 1]]
+            )
+            # Make sure that state of charge of batteries and pumped hydro
+            # plants are cyclic over the year by using the state_of_charges
+            # from the pre_market_model
+            if i == 0:
+                n.storage_units.state_of_charge_initial = (
+                    pre_market.storage_units_t.state_of_charge.iloc[-1]
+                )
+                seasonal_storage = pre_market.storage_units[
+                    pre_market.storage_units.carrier == "reservoir"
+                ].index
+
+                soc_value = pre_market.storage_units_t.state_of_charge.loc[
+                    snapshots[end - 1], seasonal_storage
+                ]
+
+                args_addition = {
+                    "pre_market_seasonal_soc": soc_value,
+                }
+
+                extra_functionality = Constraints(
+                    {**args, **args_addition}, False, apply_on="market_model"
+                ).functionality
+
+            elif i == len(starting_points) - 1:
+                if len(snapshots) > 1000:
+                    extra_functionality = Constraints(
+                        args, False, apply_on="last_market_model"
+                    ).functionality
+            else:
+                seasonal_storage = pre_market.storage_units[
+                    pre_market.storage_units.carrier == "reservoir"
+                ].index
+
+                soc_value = pre_market.storage_units_t.state_of_charge.loc[
+                    snapshots[end - 1], seasonal_storage
+                ]
+
+                args_addition = {
+                    "pre_market_seasonal_soc": soc_value,
+                }
+
+                extra_functionality = Constraints(
+                    {**args, **args_addition}, False, apply_on="market_model"
+                ).functionality
+
+        if args["method"]["formulation"] == "linopy":
+            status, condition = n.optimize(
+                sns,
+                solver_name=solver_name,
+                extra_functionality=extra_functionality,
+                assign_all_duals=True,
+                solver_options=args["solver_options"],
+            )
+
+            if status != "ok":
+                logger.warning(
+                    f"""Optimization failed with status {status}
+                    and condition {condition}"""
+                )
+                n.model.print_infeasibilities()
+                import pdb
+
+                pdb.set_trace()
+
+        else:
+            n.lopf(
+                sns,
+                solver_name=solver_name,
+                solver_options=args["solver_options"],
+                pyomo=True,
+                extra_functionality=extra_functionality,
+                formulation=args["model_formulation"],
+            )
+
+    return n
+
+
 def import_gen_from_links(network, drop_small_capacities=True):
     """
     create gas generators from links in order to not lose them when
